// src/app/common/generated-types.ts
export type Maybe<T> = T;
export type InputMaybe<T> = T;
export type Exact<T extends { [key: string]: unknown }> = {
    [K in keyof T]: T[K];
};
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & {
    [SubKey in K]?: Maybe<T[SubKey]>;
};
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & {
    [SubKey in K]: Maybe<T[SubKey]>;
};
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
    ID: string;
    String: string;
    Boolean: boolean;
    Int: number;
    Float: number;
    /** A date-time string at UTC, such as 2007-12-03T10:15:30Z, compliant with the `date-time` format outlined in section 5.6 of the RFC 3339 profile of the ISO 8601 standard for representation of dates and times using the Gregorian calendar. */
    DateTime: any;
    /** The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
    JSON: any;
    /** The `Money` scalar type represents monetary values and supports signed double-precision fractional values as specified by [IEEE 754](https://en.wikipedia.org/wiki/IEEE_floating_point). */
    Money: any;
    /** The `Upload` scalar type represents a file upload. */
    Upload: any;
};

export type ActiveOrderResult = NoActiveOrderError | Order;

export type AddPaymentToOrderResult =
    | IneligiblePaymentMethodError
    | NoActiveOrderError
    | Order
    | OrderPaymentStateError
    | OrderStateTransitionError
    | PaymentDeclinedError
    | PaymentFailedError;

export type Address = Node & {
    __typename?: "Address";
    city?: Maybe<Scalars["String"]>;
    company?: Maybe<Scalars["String"]>;
    country?: Country;
    createdAt: Scalars["DateTime"];
    customFields?: Maybe<Scalars["JSON"]>;
    defaultBillingAddress?: Maybe<Scalars["Boolean"]>;
    defaultShippingAddress?: Maybe<Scalars["Boolean"]>;
    fullName?: Maybe<Scalars["String"]>;
    id: Scalars["ID"];
    phoneNumber?: Maybe<Scalars["String"]>;
    postalCode?: Maybe<Scalars["String"]>;
    province?: Maybe<Scalars["String"]>;
    streetLine1: Scalars["String"];
    streetLine2?: Maybe<Scalars["String"]>;
    updatedAt: Scalars["DateTime"];
};
// ...
export type Country = Node &
    Region & {
        __typename?: "Country";
        code: Scalars["String"];
        createdAt: Scalars["DateTime"];
        customFields?: Maybe<Scalars["JSON"]>;
        enabled: Scalars["Boolean"];
        id: Scalars["ID"];
        languageCode: LanguageCode;
        name: Scalars["String"];
        parent?: Maybe<Region>;
        parentId?: Maybe<Scalars["ID"]>;
        translations: Array<RegionTranslation>;
        type: Scalars["String"];
        updatedAt: Scalars["DateTime"];
    };
// ...
export type Node = {
    id: Scalars["ID"];
};
// ...
export type OrderAddress = {
    __typename?: "OrderAddress";
    city?: Maybe<Scalars["String"]>;
    company?: Maybe<Scalars["String"]>;
    country?: Maybe<Scalars["String"]>;
    countryCode?: Maybe<Scalars["String"]>;
    customFields?: Maybe<Scalars["JSON"]>;
    fullName?: Maybe<Scalars["String"]>;
    phoneNumber?: Maybe<Scalars["String"]>;
    postalCode?: Maybe<Scalars["String"]>;
    province?: Maybe<Scalars["String"]>;
    streetLine1?: Maybe<Scalars["String"]>;
    streetLine2?: Maybe<Scalars["String"]>;
};
// ...
export type Region = {
    code: Scalars["String"];
    createdAt: Scalars["DateTime"];
    enabled: Scalars["Boolean"];
    id: Scalars["ID"];
    languageCode: LanguageCode;
    name: Scalars["String"];
    parent?: Maybe<Region>;
    parentId?: Maybe<Scalars["ID"]>;
    translations: Array<RegionTranslation>;
    type: Scalars["String"];
    updatedAt: Scalars["DateTime"];
};

export type RegionTranslation = {
    __typename?: "RegionTranslation";
    createdAt: Scalars["DateTime"];
    id: Scalars["ID"];
    languageCode: LanguageCode;
    name: Scalars["String"];
    updatedAt: Scalars["DateTime"];
};
// ...
export type CountryFragment = {
    __typename?: "Country";
    id: string;
    code: string;
    name: string;
    enabled: boolean;
};

export type OrderAddressFragment = {
    __typename?: "OrderAddress";
    fullName?: string;
    company?: string;
    streetLine1?: string;
    streetLine2?: string;
    city?: string;
    province?: string;
    postalCode?: string;
    country?: string;
    phoneNumber?: string;
};

export type AddressFragment = {
    __typename?: "Address";
    id: string;
    fullName?: string;
    company?: string;
    streetLine1: string;
    streetLine2?: string;
    city?: string;
    province?: string;
    postalCode?: string;
    phoneNumber?: string;
    defaultShippingAddress?: boolean;
    defaultBillingAddress?: boolean;
    country: { __typename?: "Country"; id: string; code: string; name: string };
};